requirements.txt
gunicorn
django-heroku
requests
urllib